LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-path-animation-qQ-tT-inverse.tentative-expected.txt


FAIL Path animation where coordinate modes of start and end differ (q-Q and t-T) assert_equals: expected "M -30.00 -30.00 Q 7.50 -30.00 7.50 0.00 T -30.00 30.00 Z" but got "M -30.00 -30.00 q 37.50 0.00 37.50 30.00 t -37.50 30.00 Z"
